refactor: rename ZNE mitigation step to ZNE step
