+++ core/src/oqtopus_engine_core/steps/__init__.py
@@ -6,7 +6,7 @@ from .multi_manual_step import MultiManualStep
 from .ro_error_mitigation_step import ReadoutErrorMitigationStep
 from .sse_step import SseStep
 from .tranqu_step import TranquStep
-from .zne_mitigation_step import ZneStep
+from .zne_step import ZneStep
 
 __all__ = [
     "DebugStep",
